scrapy_app/requirements.txt
scrapy==2.6.3
mysql-connector-python==8.0.28
python-dotenv==0.19.0
python-dateutil==2.8.2
requests==2.27.1
